* Made delicious-cli more user-friendly * Updated the documentation * Fixed some bug in the DB

+++ lib/delicious-cli/db.rb
@@ -35,7 +35,7 @@ class Database
   end
   
   def self.clear!
-    File.delete @@filename
+    File.delete @@filename if File.exists? @@filename
     @@posts = []
   end
